I/O functions

readmusicxml, parsemusicxml

MusicXML.extractdata

MusicXML.parsemusicxml

MusicXML.readmusicxml

MusicXML.parsemusicxml — Method

parsemusicxml(s)

Parses musicxml from a string and then extracts the data, builds all the types and stores them in proper format.

Examples

data = parsemusicxml(s)

MusicXML.readmusicxml — Method

readmusicxml(filepath)

Reads musicxml file and then extracts the data, builds all the types and stores them in proper format.

Examples

data = readmusicxml(joinpath("examples", "musescore.musicxml"))

MusicXML.extractdata — Method

extractdata(doc)

Extracts musicxml data, builds all the types and stores them in proper format.

This function is not exported. Use readmusicxml and parsemusicxml instead.

Examples

data = extractdata(doc)